% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirmatoryfactoranalysisWrapper.R
\name{confirmatoryFactorAnalysis}
\alias{confirmatoryFactorAnalysis}
\title{Confirmatory Factor Analysis}
\usage{
confirmatoryFactorAnalysis(
  data = NULL,
  version = "1",
  addFactorScoresToData = FALSE,
  addFactorScoresToDataPrefix = "FS",
  ave = FALSE,
  bartlettTest = FALSE,
  bootstrapSamples = 1000,
  ciLevel = 0.95,
  dataType = "raw",
  dependentVariablesCorrelated = TRUE,
  estimator = "default",
  exogenousCovariatesFixed = TRUE,
  exogenousLatentsCorrelated = TRUE,
  factors = list(list(indicators = list(), name = "Factor1", title = "Factor 1")),
  factorsUncorrelated = FALSE,
  fitMeasures = FALSE,
  group = list(types = "unknown", value = ""),
  htmt = FALSE,
  impliedCovarianceMatrix = FALSE,
  interceptsFixedToZero = "latent",
  invarianceTesting = "configural",
  kaiserMeyerOlkinTest = FALSE,
  latentInterceptsFixedToZero = TRUE,
  lavaanSyntax = FALSE,
  manifestInterceptsFixedToZero = FALSE,
  meanStructure = FALSE,
  misfitPlot = FALSE,
  modelIdentification = "markerVariable",
  modificationIndices = FALSE,
  modificationIndicesCutoff = 3.84,
  naAction = "fiml",
  packageMimiced = "lavaan",
  pathPlot = FALSE,
  pathPlotFontSize = 0.9,
  pathPlotMean = FALSE,
  pathPlotParameter = FALSE,
  pathPlotRotated = FALSE,
  pathPlotStandardized = FALSE,
  pathPlotVariance = TRUE,
  plotHeight = 320,
  plotWidth = 480,
  rSquared = FALSE,
  reliability = FALSE,
  residualCovarianceMatrix = FALSE,
  residualSingleIndicatorOmitted = TRUE,
  residualVariances = TRUE,
  residualsCovarying = list(),
  sampleSize = 200,
  scalingParamaters = TRUE,
  seType = "default",
  secondOrder = list(types = list(), value = list()),
  standardized = "none",
  thresholds = TRUE
)
}
\arguments{
\item{addFactorScoresToData, }{Adds the estimated factor scores as new columns to the data set
Defaults to \code{FALSE}.}
}
\description{
Confirmatory Factor Analysis
}
